RemovePartyMon

+++ include/field_specials.h
@@ -32,6 +32,6 @@ bool8 UsedPokemonCenterWarp(void);
 void ResetFanClub(void);
 bool8 ShouldShowBoxWasFullMessage(void);
 void SetPCBoxToSendMon(u8 boxId);
-bool32 CheckPartyHasSpecies(u32 givenSpecies);
+bool32 CheckPartyHasSpecies(u16 givenSpecies);
 
 #endif // GUARD_FIELD_SPECIALS_H
